+++ src/ducks/Signals/signalFormManager/signalCrudForm/formParts/TriggerProjectsSelector.js
@@ -1,10 +1,47 @@
-import TriggerProjectsSelectorWrapper, {
-  TriggerProjectsSelector
-} from './TriggerProjectsSelectorWrapper'
+import React from 'react'
+import FormikSelect from '../../../../../components/formik-santiment-ui/FormikSelect'
 
-export const ProjectsByErc20 = TriggerProjectsSelectorWrapper({
-  type: 'erc20'
-})(TriggerProjectsSelector)
-export const ProjectsAll = TriggerProjectsSelectorWrapper({ type: 'all' })(
-  TriggerProjectsSelector
-)
+export const TriggerProjectsSelector = ({
+  projects = [],
+  onChange,
+  metaFormSettings,
+  target,
+  setFieldValue
+}) => {
+  const { ethAddress, target: defaultAsset } = metaFormSettings
+
+  const options = projects.map(asset => ({
+    label: asset.slug,
+    value: asset.slug
+  }))
+
+  if (
+    target &&
+    target.value &&
+    options.length > 0 &&
+    !options.find(p => p.value === target.value)
+  ) {
+    setFieldValue('target', '')
+  }
+
+  return (
+    <FormikSelect
+      name='target'
+      isDisabled={defaultAsset.isDisabled}
+      defaultValue={defaultAsset.value.value}
+      placeholder='Pick an asset'
+      required
+      options={options}
+      onChange={newAsset => {
+        if (ethAddress) {
+          if (metaFormSettings.target.value.value === newAsset.value) {
+            setFieldValue('ethAddress', ethAddress)
+          } else {
+            setFieldValue('ethAddress', '')
+          }
+        }
+        onChange && onChange(newAsset)
+      }}
+    />
+  )
+}
